+++ src/main/kotlin/com/atlassian/performance/tools/report/api/ShiftedDistributionRegressionTest.kt
@@ -14,6 +14,7 @@ import org.apache.commons.math3.stat.descriptive.rank.Median
  * @param mwAlpha Mann-Whitney significance level
  * @param ksAlpha Kolmogorov-Smirnov significance level
  */
+@Deprecated("Use DistributionComparator instead")
 class ShiftedDistributionRegressionTest(
     private val baseline: DoubleArray,
     private val experiment: DoubleArray,
@@ -57,7 +58,12 @@ class ShiftedDistributionRegressionTest(
     }
 
     internal fun overcomesTolerance(tolerance: Double): Boolean {
-        return isExperimentRegressed(tolerance) || isExperimentImproved(tolerance)
+        val isExperimentRegressed = isExperimentRegressed(tolerance)
+        val isExperimentImproved  = isExperimentImproved(tolerance)
+        if (isExperimentImproved && isExperimentRegressed) {
+            throw IllegalArgumentException("Experiment can't be both regressed and improved at the same time")
+        }
+        return isExperimentRegressed || isExperimentImproved
     }
 
     /**